- workers should be much more robust - bug fix for download world url (getting server.ip_address instead of server.remote.ip_address) - bug fix for server log messages > 255 chars

+++ app/workers/wait_for_snapshotting_server_worker.rb
@@ -1,44 +1,42 @@
 class WaitForSnapshottingServerWorker
   include Sidekiq::Worker
-  sidekiq_options retry: 4
-  sidekiq_retry_in do |count|
-    4
-  end
+  sidekiq_options retry: 0
 
-  sidekiq_retries_exhausted do |msg|
-    args = msg['args']
-    server = Server.find(args[0])
-    server.minecraft.log("Background job waiting for snapshotting server died: #{msg['error_message']}")
-  end
-
-  def perform(server_id, digital_ocean_snapshot_action_id)
+  def perform(server_id, digital_ocean_action_id, times = 0)
     server = Server.find(server_id)
+    if times > 32
+      server.minecraft.log("Still waiting for Digital Ocean server to snapshot, tried #{times} times")
+    elsif times > 64
+      server.minecraft.log('Digital Ocean took too long to snapshot server. Aborting')
+      server.reset_partial
+      return
+    end
     if !server.remote.exists?
       server.minecraft.log('Error stopping server; remote_id is nil. Aborting')
-      server.reset
+      server.reset_partial
       return
     end
     if server.remote.error?
       server.minecraft.log("Error communicating with Digital Ocean while stopping server; they responded with #{server.remote.error}. Aborting")
-      server.reset
+      server.reset_partial
       return
     end
-    event = DigitalOcean::Action.new(server.remote_id, digital_ocean_snapshot_action_id, server.minecraft.user)
+    event = DigitalOcean::Action.new(server.remote_id, digital_ocean_action_id, server.minecraft.user)
     if event.error?
-      server.minecraft.log("Error with Digital Ocean snapshotting server action #{digital_ocean_snapshot_action_id}; they responded with #{event.show}")
-      error = server.remote.shutdown
-      if error
-        server.minecraft.log("Error shutting down server on Digital Ocean; #{error}. Aborting")
-        server.reset_partial
-        return
-      end
-      WaitForStoppingServerWorker.perform_in(4.seconds, server_id)
+      server.minecraft.log("Error with Digital Ocean snapshot server action #{digital_ocean_action_id}; they responded with #{event.show}. Aborting")
+      server.reset_partial
       return
     elsif !event.done?
-      WaitForSnapshottingServerWorker.perform_in(4.seconds, server_id, digital_ocean_snapshot_action_id)
+      WaitForSnapshottingServerWorker.perform_in(4.seconds, server_id, digital_ocean_action_id, times + 1)
       return
     end
-    server.update_columns(do_saved_snapshot_id: server.remote.latest_snapshot_id)
+    snapshot_id = server.remote.latest_snapshot_id
+    if snapshot_id.nil?
+      server.minecraft.log('Finished snapshotting server on Digital Ocean, but unable to get latest snapshot id. Aborting')
+      server.reset_partial
+      return
+    end
+    server.update_columns(do_saved_snapshot_id: snapshot_id)
     error = server.remote.destroy
     if error
       server.minecraft.log("Error destroying server on Digital Ocean (has been snapshotted and saved); #{error}")
@@ -48,6 +46,7 @@ class WaitForSnapshottingServerWorker
     logger.info "Record in #{self.class} not found #{e.message}"
   rescue => e
     server.minecraft.log("Background job waiting for snapshotting server failed: #{e}")
+    server.reset_partial
     raise
   end
